requirement.txt
flasgger
Flask
Flask-RESTful
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
pymongo
python-dateutil==2.9.0.post0
python-dotenv
pandas
google-cloud-secret-manager
openpyxl
cors
gunicorn
firebase_admin
flask_cors